requirements.txt
flask==2.0.1
flask-cors==3.0.10
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
numpy==1.20.3
scikit-learn==0.24.2
python-dotenv==0.19.0
openai==0.28.0
gunicorn==20.1.0